requirements.txt
openai
pysrt
python-dotenv
tqdm
